move ultiSnips editSnippets command to provider to make room for masscode
